backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
lightgbm==4.5.0
torch==2.4.1
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3
